requirements.txt
Flask==2.3.2
python-docx==0.8.11
PyYAML==6.0.1
Werkzeug==2.3.6
reportlab==4.0.4
jinja2==3.1.2
markupsafe==2.1.3